requirements.txt
bs4
requests
urllib
python-dotenv